resolve conflicts 3
