--- src/compression/dwa/csc.rs
@@ -1,24 +1,0 @@
-// Y'CbCr -> R'G'B' inverse conversion for DWA, using the modified 709
-// coefficients OpenEXR's DWA encoder uses (from OpenEXRCore
-// internal_dwa_simd.h).
-
-/// Input comp0/1/2 are Y, RY, BY; output is R, G, B.
-#[inline]
-pub fn csc709_inverse(comp0: f32, comp1: f32, comp2: f32) -> (f32, f32, f32) {
-    let r = comp0 + 1.5747 * comp2;
-    let g = comp0 - 0.1873 * comp1 - 0.4682 * comp2;
-    let b = comp0 + 1.8556 * comp1;
-    (r, g, b)
-}
-
-/// R'G'B' -> Y'CbCr forward conversion for DWA. The component order matches
-/// OpenEXR's channel-group storage: Y, BY, RY.
-#[inline]
-pub fn csc709_forward(r: f32, g: f32, b: f32) -> (f32, f32, f32) {
-    // OpenEXR uses a modified 709 transform with a zero-centered chroma
-    // representation instead of the usual 0.5 offset.
-    let y = 0.2126 * r + 0.7152 * g + 0.0722 * b;
-    let by = (b - y) / 1.8556;
-    let ry = (r - y) / 1.5747;
-    (y, by, ry)
-}
